+++ berserk/session.py
@@ -58,7 +58,7 @@ class Requestor(Generic[T]):
         converter: Converter[Any] = utils.noop,
         **kwargs: Any,
     ) -> Any | Iterator[Any]:
-        """Make a request for a resource in a paticular format.
+        """Make a request for a resource in a particular format.
 
         :param method: HTTP verb
         :param path: the URL suffix