+++ berserk/types/__init__.py
@@ -1,12 +1,21 @@
 from __future__ import annotations
 
-from .account import AccountInformation, Perf, Preferences, Profile, StreamerInfo
+
+from .account import (
+    AccountInformation,
+    Perf,
+    Preferences,
+    Profile,
+    StreamerInfo,
+    Timeline,
+)
 from .broadcast import (
     BroadcastPlayer,
     BroadcastTop,
     PaginatedBroadcasts,
     BroadcastsByUser,
 )
+
 from .bulk_pairings import BulkPairing, BulkPairingGame
 from .challenges import ChallengeJson
 from .common import ClockConfig, ExternalEngine, LightUser, OnlineLightUser, VariantKey
@@ -52,6 +61,7 @@ __all__ = [
     "SwissInfo",
     "SwissResult",
     "Team",
-    "VariantKey",
+    "Timeline",
     "TVFeed",
+    "VariantKey",
 ]