+++ berserk/types/account.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from typing_extensions import TypedDict, TypeAlias, Literal
+from typing_extensions import TypedDict, TypeAlias, Literal, NotRequired, Union
+from .common import LightUser, PerfType, Title
 
 
 class Perf(TypedDict):
@@ -10,16 +11,16 @@ class Perf(TypedDict):
     rating: int
     rd: int
     prog: int
-    prov: bool
+    prov: NotRequired[bool]
 
 
 class Profile(TypedDict):
     """Public profile of an account."""
 
-    flag: str
-    location: str
+    flag: NotRequired[str]
+    location: NotRequired[str]
     bio: str
-    realName: str
+    realName: NotRequired[str]
     fideRating: int
     uscfRating: int
     ecfRating: int
@@ -43,24 +44,34 @@ class StreamerInfo(TypedDict):
 class AccountInformation(TypedDict):
     """Information about an account."""
 
+    # Same as Light user except name <-> username
+
+    # The id of the user
     id: str
+    # The name of the user
     username: str
+    # The title of the user
+    title: NotRequired[Title]
+    # The flair of the user
+    flair: NotRequired[str]
+    # The patron of the user
+    patron: NotRequired[bool]
+    # The patron color of the user
+    patronColor: NotRequired[int]
+
     perfs: dict[str, Perf]
-    flair: str
     createdAt: datetime
-    disabled: bool
-    tosViolation: bool
+    disabled: NotRequired[bool]
+    tosViolation: NotRequired[bool]
     profile: Profile
     seenAt: datetime
-    patron: bool
-    verified: bool
+    verified: NotRequired[bool]
     playTime: PlayTime
-    title: str
     url: str
-    playing: str
+    playing: NotRequired[str]
     count: dict[str, int]
-    streaming: bool
-    streamer: dict[str, StreamerInfo]
+    streaming: NotRequired[bool]
+    streamer: NotRequired[dict[str, StreamerInfo]]
     followable: bool
     following: bool
     blocking: bool
@@ -120,3 +131,182 @@ class UserPreferences(TypedDict, total=False):
 class Preferences(TypedDict):
     prefs: UserPreferences
     language: str
+
+
+# from lila/modules/core/src/main/timeline.scala
+
+
+class TimelineEntry(TypedDict):
+    date: datetime
+
+
+# ---
+
+
+class FollowEntry(TimelineEntry):
+    data: FollowData
+    type: Literal["follow"]
+
+
+class FollowData(TypedDict):
+    u1: str
+    u2: str
+
+
+# ---
+
+
+class TeamJoinEntry(TimelineEntry):
+    data: TeamJoinData
+    type: Literal["team-join"]
+
+
+class TeamJoinData(TypedDict):
+    userId: str
+    teamId: str
+
+
+# ---
+
+
+class TeamCreateEntry(TimelineEntry):
+    data: TeamCreateData
+    type: Literal["team-create"]
+
+
+class TeamCreateData(TypedDict):
+    userId: str
+    teamId: str
+
+
+# ---
+
+
+class ForumPostEntry(TimelineEntry):
+    data: ForumPostData
+    type: Literal["forum-post"]
+
+
+class ForumPostData(TypedDict):
+    userId: str
+    topicId: str
+    topicName: str
+    postId: str
+
+
+# ---
+
+
+class UblogPostEntry(TimelineEntry):
+    data: UblogPostData
+    type: Literal["ublog-post", "ublog-post-like"]
+
+
+class UblogPostData(TypedDict):
+    userId: str
+    id: str
+    title: str
+
+
+# ---
+
+
+class TourJoinEntry(TimelineEntry):
+    data: TourJoinData
+    type: Literal["tour-join"]
+
+
+class TourJoinData(TypedDict):
+    userId: str
+    tourId: str
+    tourName: str
+
+
+# ---
+
+
+class GameEndEntry(TimelineEntry):
+    data: GameEndData
+    type: Literal["game-end"]
+
+
+class GameEndData(TypedDict):
+    fullId: str
+    opponent: NotRequired[str]
+    win: NotRequired[bool]
+    perf: PerfType
+
+
+# ---
+
+
+class SimulEntry(TimelineEntry):
+    data: SimulData
+    type: Literal["simul-create", "simul-join"]
+
+
+class SimulData(TypedDict):
+    userId: str
+    simulId: str
+    simulName: str
+
+
+# ---
+
+
+class StudyLikeEntry(TimelineEntry):
+    data: StudyLikeData
+    type: Literal["study-like"]
+
+
+class StudyLikeData(TypedDict):
+    userId: str
+    studyId: str
+    studyName: str
+
+
+# ---
+
+
+class PlanStartEntry(TimelineEntry):
+    data: PlanStartData
+    type: Literal["plan-start"]
+
+
+class PlanStartData(TypedDict):
+    userId: str
+
+
+# ---
+
+
+class PlanRenewEntry(TimelineEntry):
+    data: PlanRenewData
+    type: Literal["plan-renew"]
+
+
+class PlanRenewData(TypedDict):
+    userId: str
+    months: int
+
+
+# ---
+
+TimelineEntries = Union[
+    FollowEntry,
+    TeamJoinEntry,
+    TeamCreateEntry,
+    ForumPostEntry,
+    UblogPostEntry,
+    TourJoinEntry,
+    GameEndEntry,
+    StudyLikeEntry,
+    SimulEntry,
+    PlanStartEntry,
+    PlanRenewEntry,
+]
+
+
+class Timeline(TypedDict):
+    entries: list[TimelineEntries]
+    users: dict[str, LightUser]