+++ integration/local.py
@@ -201,6 +201,7 @@ def integration_test(python_version: str, watch: bool) -> None:
                 BDIT_NETWORK,
                 BDIT_APP_IMAGE,
             ],
+            check=False,
         )
         log.info("App container finished successfully.")